Move logger to core
set log handler based on environment
